# Conflicts:
#	src/main/java/poomasi/global/error/BusinessError.java

+++ src/main/java/poomasi/domain/farm/entity/Farm.java
@@ -57,7 +57,7 @@ public class Farm {
     private FarmStatus status = FarmStatus.OPEN;
 
     @Comment("체험 비용")
-    private Long experiencePrice;
+    private int experiencePrice;
 
     @Comment("팀 최대 인원")
     private Integer maxCapacity;
@@ -85,7 +85,7 @@ public class Farm {
     private OrderedFarm orderedFarm;
 
     @Builder
-    public Farm(Long id, String name, Long ownerId, String address, String addressDetail, Double latitude, Double longitude, String description, Long experiencePrice, Integer maxCapacity, Integer maxReservation) {
+    public Farm(Long id, String name, Long ownerId, String address, String addressDetail, Double latitude, Double longitude, String description, int experiencePrice, Integer maxCapacity, Integer maxReservation) {
         this.id = id;
         this.name = name;
         this.ownerId = ownerId;
@@ -109,7 +109,7 @@ public class Farm {
         return this;
     }
 
-    public void updateExpPrice(Long expPrice) {
+    public void updateExpPrice(int expPrice) {
         this.experiencePrice = expPrice;
     }
